common util function
