+++ src/include/duckdb/function/scalar/string_common.hpp
@@ -18,7 +18,7 @@ string_t SubstringGrapheme(Vector &result, string_t input, int64_t offset, int64
 bool SubstringInSupportedRange(int64_t offset, int64_t length);
 unique_ptr<BaseStatistics> PropagateStringSliceStats(FunctionStatisticsInput &input, idx_t start_character_index,
                                                      optional_idx character_count);
-//! Zonemap pruning for `prefix(s, constant)`, shared with starts_with / ^@
+//! Common util zonemap pruning for `prefix(s, constant)`.
 FilterPropagateResult PrefixFilterPrune(const FunctionStatisticsPruneInput &input);
 
 ScalarFunction GetStringContains();
